Add projection.{pre,post}clip.
This allows arbitrary implementations of spherical clipping and pixel clipping
to be specified on a projection. We’re still not achieving the generality of a
composable projection pipeline, but at least we can plug in new implementations,
such as the forthcoming spherical polygon clipping (#108).
