+++ src/projection/identity.js
@@ -1,4 +1,4 @@
-import {clipExtent} from "../clip/extent";
+import clipRectangle from "../clip/rectangle";
 import identity from "../identity";
 import {transformer} from "../transform";
 import {fitExtent, fitSize} from "./fit";
@@ -13,7 +13,8 @@ function scaleTranslate(kx, ky, tx, ty) {
 
 export default function() {
   var k = 1, tx = 0, ty = 0, sx = 1, sy = 1, transform = identity, // scale, translate and reflect
-      x0 = null, y0, x1, y1, clip = identity, // clip extent
+      x0 = null, y0, x1, y1, // clip extent
+      postclip = identity,
       cache,
       cacheStream,
       projection;
@@ -25,10 +26,13 @@ export default function() {
 
   return projection = {
     stream: function(stream) {
-      return cache && cacheStream === stream ? cache : cache = transform(clip(cacheStream = stream));
+      return cache && cacheStream === stream ? cache : cache = transform(postclip(cacheStream = stream));
+    },
+    postclip: function(_) {
+      return arguments.length ? (postclip = _, x0 = y0 = x1 = y1 = null, reset()) : postclip;
     },
     clipExtent: function(_) {
-      return arguments.length ? (clip = _ == null ? (x0 = y0 = x1 = y1 = null, identity) : clipExtent(x0 = +_[0][0], y0 = +_[0][1], x1 = +_[1][0], y1 = +_[1][1]), reset()) : x0 == null ? null : [[x0, y0], [x1, y1]];
+      return arguments.length ? (postclip = _ == null ? (x0 = y0 = x1 = y1 = null, identity) : clipRectangle(x0 = +_[0][0], y0 = +_[0][1], x1 = +_[1][0], y1 = +_[1][1]), reset()) : x0 == null ? null : [[x0, y0], [x1, y1]];
     },
     scale: function(_) {
       return arguments.length ? (transform = scaleTranslate((k = +_) * sx, k * sy, tx, ty), reset()) : k;
